# PPL Coalesce Function Enhancement for Calcite - Advanced Type Handling and Field Support