--- a/core/src/main/java/org/opensearch/sql/calcite/CalciteRexNodeVisitor.java
+++ b/core/src/main/java/org/opensearch/sql/calcite/CalciteRexNodeVisitor.java
@@ -541,7 +541,8 @@ public class CalciteRexNodeVisitor extends AbstractNodeVisitor<RexNode, CalciteP
                 new UnsupportedOperationException(
                     "Unexpected window function: " + windowFunction.getFuncName()));
   }
-
+  
+  /** extract the expression of Alias from a node */
   private RexNode extractRexNodeFromAlias(RexNode node) {
     requireNonNull(node);
     if (node.getKind() == AS) {