… core/src/main/java/org/opensearch/sql/calcite/utils/UserDefinedFunctionUtils.java
… core/src/main/java/org/opensearch/sql/expression/function/PPLBuiltinOperators.java
… core/src/main/java/org/opensearch/sql/expression/function/PPLFuncImpTable.java
… core/src/main/java/org/opensearch/sql/expression/function/udf/condition/EnhancedCoalesceFunction.java
… docs/user/ppl/functions/condition.rst
… integ-test/src/test/java/org/opensearch/sql/calcite/remote/CalcitePPLEnhancedCoalesceIT.java
… integ-test/src/test/java/org/opensearch/sql/security/CrossClusterCoalesceIT.java
… ppl/src/test/java/org/opensearch/sql/ppl/calcite/CalcitePPLEnhancedCoalesceTest.java
… ppl/src/test/java/org/opensearch/sql/ppl/calcite/CalcitePPLFunctionTypeTest.java