# PPL Coalesce Function Enhancement for Calcite - Advanced Type Handling and Field Support

## Description

This PR implements significant enhancements to the existing PPL `coalesce` function in Calcite to address current limitations with strict type checking and field requirements. The enhancement transforms coalesce from a restrictive function that requires same-type arguments and existing fields to a flexible function that handles mixed data types, non-existent fields, and empty values gracefully.

### Features Implemented

**✅ Feature 1: Automatic Type Coercion**
- Removed strict same-type requirement for all arguments
- Implemented intelligent type coercion for mixed data types
- Example: `source=data | eval result = coalesce(field1, 123, "fallback")`

**✅ Feature 2: Non-existent Field Support**
- Added support for dummy/non-existent fields in coalesce arguments
- Graceful handling of missing fields without runtime errors
- Example: `source=firewall | eval src_ip = coalesce(src_ip, sourceip, dummy_field, "unknown")`

**✅ Feature 3: Empty Field Handling**
- Support for empty strings ("") and whitespace strings (" ") as valid values
- Empty strings are treated as valid non-null values, not skipped
- Example: `source=accounts | eval result = coalesce("", firstname)` returns ""

## Technical Implementation

### Calcite Changes
- Updated coalesce function to support mixed data types with automatic type coercion
- Enhanced field resolution to handle non-existent fields without errors
- Modified value processing to treat empty strings as valid non-null values

## Usage Examples

```ppl
# Feature 1: Automatic Type Coercion
source=data | eval result = coalesce(field1, 123, "fallback")
source=logs | eval value = coalesce(numeric_field, "0", 999)

# Feature 2: Non-existent Field Support  
source=firewall | eval src_ip = coalesce(src_ip, sourceip, dummy_field, "unknown")
source=logs | eval ip = coalesce(clientip, ipaddress, non_existent_field, "127.0.0.1")

# Feature 3: Empty Field Handling
source=accounts | eval result = coalesce("", firstname)        # Returns ""
source=accounts | eval result = coalesce(" ", firstname)       # Returns " "
source=accounts | eval result = coalesce(firstname, "")        # Returns firstname or ""

# Combined Features Example
source=firewall | eval src_ip = coalesce(src_ip, sourceip, dummy_field, "unknown")  # Uses all 3 features
```
## Testing

### Testing Plan
- **Type Coercion**: Test mixed data type handling and automatic conversions
- **Non-existent Fields**: Test graceful handling of missing/dummy fields
- **Empty Values**: Test empty string and whitespace processing
- **Backward Compatibility**: Verify existing queries continue to work
- **Performance**: Validate no significant performance degradation

## Performance Impact

Type coercion and field resolution changes may have minimal performance overhead. Performance testing will be conducted with big5 dataset to ensure acceptable impact.

## Breaking Changes

**None** - This PR maintains full backward compatibility while adding enhanced functionality.

## Functionality Improvements

| Feature | Before Enhancement | After Enhancement |
|---------|-------------------|-------------------|
| Mixed Types | ❌ Runtime Error | ✅ Automatic Coercion |
| Non-existent Fields | ❌ Runtime Error | ✅ Graceful Handling |
| Empty Strings | ❌ Not Supported | ✅ Treated as Valid |
| Type Safety | ✅ Strict Checking | ✅ Intelligent Coercion |

## Documentation

PPL command documentation is updated in this PR to reflect the new coalesce behavior including type coercion, non-existent field handling, and empty value support.

## Best Practices

- **Field Ordering**: Place most specific/reliable fields first in coalesce arguments
- **Type Consistency**: While mixed types are supported, prefer consistent types when possible
- **Fallback Values**: Always include a literal fallback value as the last argument
- **Performance**: Use coalesce early in the query pipeline for optimal performance

## BIG5 COALESCE PERFORMANCE TEST
Dataset: big5
Test Date: Tue Aug 19 2025

### Baseline:
`source=big5 | head <size>`
head 10: Avg = 19ms, P90 = 25ms
head 100: Avg = 32ms, P90 = 38ms
head 1K: Avg = 123ms, P90 = 124ms
head 10K: Avg = 1170ms, P90 = 1222ms
head 100K: Avg = 1109ms, P90 = 1134ms

`source=big5 | eval result = coalesce(`host.name`, 'unknown') | head <size>`
head 10: Avg = 37ms, P90 = 40ms
head 100: Avg = 40ms, P90 = 44ms
head 1K: Avg = 147ms, P90 = 164ms
head 10K: Avg = 1167ms, P90 = 1177ms
head 100K: Avg = 1160ms, P90 = 1196ms

`source=big5 | eval result = coalesce(`host.name`, `agent.name`, 'unknown') | head <size>`
head 10: Avg = 29ms, P90 = 33ms
head 100: Avg = 46ms, P90 = 49ms
head 1K: Avg = 157ms, P90 = 199ms
head 10K: Avg = 1194ms, P90 = 1221ms
head 100K: Avg = 1166ms, P90 = 1203ms

`source=big5 | eval result = coalesce(`host.name`, 'unknown') | fields result, `host.name` | head <size>`
head 10: Avg = 25ms, P90 = 27ms
head 100: Avg = 28ms, P90 = 28ms
head 1K: Avg = 56ms, P90 = 57ms
head 10K: Avg = 367ms, P90 = 391ms
head 100K: Avg = 347ms, P90 = 354ms

### Feature testing:
`source=big5 | eval result = coalesce(`host.name`, 123, 'fallback') | head <size>`
head 10: Avg = 33ms, P90 = 37ms
head 100: Avg = 48ms, P90 = 52ms
head 1K: Avg = 166ms, P90 = 196ms
head 10K: Avg = 1138ms, P90 = 1163ms
head 100K: Avg = 1181ms, P90 = 1206ms

`source=big5 | eval result = coalesce(`host.name`, dummy_field, 'unknown') | head <size>`
head 10: Avg = 32ms, P90 = 38ms
head 100: Avg = 40ms, P90 = 43ms
head 1K: Avg = 135ms, P90 = 137ms
head 10K: Avg = 1193ms, P90 = 1239ms
head 100K: Avg = 1185ms, P90 = 1247ms

`source=big5 | eval result = coalesce('', `host.name`) | head <size>`
head 10: Avg = 45ms, P90 = 62ms
head 100: Avg = 37ms, P90 = 39ms
head 1K: Avg = 148ms, P90 = 172ms
head 10K: Avg = 1177ms, P90 = 1177ms
head 100K: Avg = 1143ms, P90 = 1152ms

## Related Issues
Adresses #[4005](https://github.com/opensearch-project/sql/issues/4005)
<!-- List any other related issues here -->

### Check List
- [✅] New functionality includes testing.
- [✅] New functionality has been documented.
- [✅] New functionality has javadoc added.
- [✅] New functionality has a user manual doc added.
- [ ] API changes companion pull request [created](https://github.com/opensearch-project/opensearch-api-specification/blob/main/DEVELOPER_GUIDE.md).
- [✅] Commits are signed per the DCO using `--signoff`.
- [✅] Public documentation issue/PR [created](https://github.com/opensearch-project/sql/issues/4005).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/sql/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).