@aalva500-prog 2025-08-19
Hi @dai-chen, thanks for raising these points! Please find below my answers, as well as implementation and design decision:

### 1) Is Automatic Type Coercion applied to all other functions as well?
As far as I'm concerned, no, automatic type coercion is NOT applied to all other functions. I'm not aware if the enhancements for the coalesce function are going to be applied to other functions. We may need to verify and confirm with the team.

Currently, the below is what we have:

**For condition functions:**
- COALESCE & IFNULL: Get Calcite's automatic type coercion
- IS_NULL, IS_NOT_NULL, ISPRESENT: Don't need type coercion (single operand)
- EARLIEST, LATEST: Custom UDFs without automatic coercion
- IF, CASE: Use manual type checking

**For `eval` related functions (Legacy engine):**
- No automatic type coercion - require strict type matching
- Use manual type checking and casting
- Implemented in the legacy expression evaluation system

**Other functions WITHOUT Automatic Type Coercion:**
- UDF Functions: Use strict type checking via PPLTypeChecker
- Math Functions: Implement manual type handling (e.g., DivideFunction)
- Custom Functions: Require explicit type matching

### 2) Besides UDF, did we consider other approaches? e.g., handle nonexistent or empty field in Analyzer layer?
Yes, other approaches were considered:
1. Analyzer Layer Handling (like eval functions):
    - Yes, this could handle non-existent fields in ExpressionAnalyzer.visitIdentifier()
    - However, it would require changes to core analysis pipeline
    - Also, it would work for legacy engine but not leverage Calcite's capabilities

2. Native Calcite Integration: Direct use of SqlStdOperatorTable.COALESCE
    - This is the standard Calcite operator
    - Type inference occurs in Calcite's leastRestrictive() method - Called by Calcite internally
    - Error thrown by Calcite's type system - When it cannot find a common type
       ```
        Cannot infer return type for COALESCE; operand types: [VARCHAR, INTEGER, VARCHAR]
       ```

### 3) Why UDF Approach Was Chosen:
- Separation of Concerns: Field handling is done at the visitor level, not analyzer level. Type coercion is done at Calcite level
- Calcite Integration: Leverages Calcite's engine. UDF bridges PPL syntax to Calcite's native operators seamlessly
- Type System Compatibility: Leverages Calcite's sophisticated type coercion automatically
- Minimal Impact: Doesn't require changes to core analysis pipeline
- Performance: Calcite's optimized type handling vs manual eval function logic
- Future-proof: Works with Calcite engine evolution

**The implementation shows a hybrid approach:**
- Field resolution handled in CalciteRexNodeVisitor (visitor layer)
- Type coercion delegated to Calcite's native COALESCE operator
- UDF wrapper provides the bridge between PPL and Calcite

### 4) Why Creating Our Own Enhanced Coalesce Function:

The data type evaluation happens entirely within Calcite's core libraries, not in this project's codebase. This project only calls SqlStdOperatorTable.COALESCE, and Calcite handles the rest.

Currently COALESCE Data Type Evaluation in this Project works in the following way:

1. Type Checking:
The following error  comes from Calcite's core libraries:
  ```
  Cannot infer return type for COALESCE; operand types: [VARCHAR, INTEGER, VARCHAR]
  ```

The Calcite libraries are external dependencies that contain the rigid type system causing the mixed-type issues.

2. How COALESCE Type Evaluation Works and Where It Happens:
    * Calcite's SqlStdOperatorTable.COALESCE - This is the standard Calcite operator. 
    * Type inference occurs in Calcite's leastRestrictive() method - Called by Calcite internally
    * Error thrown by Calcite's type system - When it cannot find a common type

3. Currently, in this Project's Code:
    - PPLFuncImpTable.java
    ```
     return builder.makeCall(SqlStdOperatorTable.COALESCE, processedArgs);
     ```

    - ExtendedRexBuilder.java:
    ```
    public RelDataType commonType(RexNode... nodes) {
         return this.getTypeFactory()
             .leastRestrictive(Arrays.stream(nodes).map(RexNode::getType).toList());
    }
    ```

4. Calcite Libraries Location:
    * Gradle Dependencies (build.gradle):
        - api('org.apache.calcite:calcite-core:1.38.0')
        - api 'org.apache.calcite:calcite-linq4j:1.38.0'

   * Physical JAR Files:
      ```
       ~/.gradle/caches/modules-2/files-2.1/org.apache.calcite/calcite-core/1.38.0/
      ├── calcite-core-1.38.0.jar                    # Compiled classes
      ├── calcite-core-1.38.0-sources.jar           # Source code
     └── calcite-core-1.38.0-javadoc.jar           # Documentation
      ```

5. COALESCE Implementation Files:
    * SqlCoalesceFunction.java - Main COALESCE function definition
    * RexImpTable$CoalesceImplementor.class - Runtime implementation

6. Key Code in SqlCoalesceFunction.java:
    - SqlCoalesceFunction
      ```
      public SqlCoalesceFunction() {
        super("COALESCE",
            SqlKind.COALESCE,
            ReturnTypes.LEAST_RESTRICTIVE.andThen(SqlTypeTransforms.LEAST_NULLABLE), // ← TYPE INFERENCE HERE
      null,
            OperandTypes.SAME_VARIADIC,  // ← REQUIRES SAME TYPES
            SqlFunctionCategory.SYSTEM);
      }
      ```
    
     - COALESCE Rewrite Logic:
        ```
          @Override public SqlNode rewriteCall(SqlValidator validator, SqlCall call) {
              // COALESCE gets rewritten to CASE statement:
          // COALESCE(a, b, c) becomes:
          // CASE WHEN a IS NOT NULL THEN a
          //      WHEN b IS NOT NULL THEN b  
          //      ELSE c END
          }
        ```

    - The Problem:
        * ReturnTypes.LEAST_RESTRICTIVE - Calls leastRestrictive() method
        * OperandTypes.SAME_VARIADIC - - Expects all operands to be the same type
        * When mixed types like [VARCHAR, INTEGER, VARCHAR]are passed:
            - leastRestrictive() returns null
            - Calcite throws "Cannot infer return type for COALESCE"

 6. The Solution Needed:
      To fix this, we would need to either:
      * Modify Calcite's type inference system (not feasible)
      * Register our custom function instead of using SqlStdOperatorTable.COALESCE